Final reorganization of image viewer components
