--- kmaps-app/src/views/Images/images.sass
@@ -1,171 +1,0 @@
-.c-image__context
-  padding-left: 0
-  padding-right: 0
-
-.c-image__viewer
-  background-color: #222
-  border-bottom: 10px solid #b49c59
-  width: 100%
-
-  .c-image__viewer-row
-    padding-top: 30px
-    margin-bottom: 10px
-    text-align: center
-
-    .page-control span
-      display: block
-      position: absolute
-      margin-top: 45%
-      color: #555
-      font-size: 3rem
-      cursor: pointer
-
-      @media (max-width: 1440px)
-        margin-top: 110%
-
-
-    .page-control span:hover
-      color: #DDD
-
-    .page-control span:hover
-      font-weight: bolder
-
-    .page-control.before span
-      margin-left: 89%
-      margin-right: 0
-
-      @media (max-width: 1440px)
-        margin-left: 78%
-
-  .react-iiif-viewer
-    .fullscreen
-      position: absolute
-      left: 47%
-      top: -40px
-
-    button
-      font-weight: bold
-      background: transparent
-      color: white
-
-    button svg
-      display: none
-
-    button span
-      color: #999
-      font-size: 1.75rem
-
-    button span:hover
-      color: #DDD
-
-    .zoom button
-      border: none
-      border-bottom: none!important
-
-  .c-image__caption
-    color: white
-    text-align: center
-    letter-spacing: .15rem
-
-  .c-image__title
-    margin-top: 0px
-    color: white
-    letter-spacing: .15rem
-
-    span
-      color: white
-      padding-right: 1rem
-      vertical-align: sub
-
-  .c-image__byline span
-    margin: 0 6px
-
-  .c-image__carousel
-    width: calc(100% - 24px)
-    overflow-x: auto
-    overflow-y: hidden
-    white-space: nowrap
-    margin: 20px 12px 0 12px
-
-    .thumb
-      display: inline-block
-      height: 80px
-      margin-bottom: 4px
-      cursor: pointer
-      border-radius: 3px
-      border: 2px solid #000
-
-      img
-        height: 100%
-
-    .thumb.current
-      border-color: red
-
-.c-image__metadata
-
-  .l-top.row .l-first.col
-    padding-left: 5rem
-
-  h5.c-image__colhead
-    font-size: 120%
-    letter-spacing: 0.0475rem
-    text-transform: uppercase
-    margin-top: 2rem
-
-  .u-data.row
-    line-height: 3rem
-
-  span[class*="u-icon"]
-    padding-right: 4px
-
-  .u-visibility
-    &:before
-      margin-left: 4px
-      content: '('
-    &:after
-      content: ')'
-
-  .l-meta.row
-    min-height: 400px
-    margin-top: 2rem
-    border-top: 4px solid #b49c59
-    color: #000000
-    background-color: white
-    padding: 2rem 8rem
-
-  .c-image__metatitle
-    color: #777
-    letter-spacing: .1rem
-
-  .c-image__caprow
-    margin-bottom: 1rem
-
-  .o-desc__title
-    font-weight: bold
-
-  .o-desc__byline
-    text-align: right
-    margin-top: -2.5rem
-    font-size: 90%
-    font-style: italic
-    &:before
-      content: 'by '
-
-
-.u-bo
-  font-family: Jomolhari, 'Tibetan Machine Uni', sans
-  font-size: 150%
-
-.u-label
-  font-family: "Open Sans","Segoe UI",Tahoma,sans-serif,shanticon
-  font-weight: 400
-  letter-spacing: 0.0475rem
-  text-transform: uppercase
-
-.u-value
-  font-family: "EB Garamond",serif
-  font-size: 125%
-
-.u-icon__copyright:before
-  content: '© '
-  font-size: 1.8rem
